Fix bug on previous timestamp data type
Variable 'previous_timestamp_'  in FusionEKF.h has been changed type to
avoid overflow problems

+++ project_6_extended_kalman_filter/src/FusionEKF.h
@@ -31,7 +31,7 @@ private:
   bool is_initialized_;
 
   // previous timestamp
-  long previous_timestamp_;
+  long long previous_timestamp_;
 
   // tool object used to compute Jacobian and RMSE
   Tools tools;
